separate issue and redeem in folders and use new From input
