+++ src/pages/bridge/Redeem/index.tsx
@@ -1,29 +1,30 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { VoidFn } from '@polkadot/api-base/types';
 import { SubmittableExtrinsic } from '@polkadot/api/promise/types';
 import Big from 'big.js';
 import _ from 'lodash';
-import { DateTime } from 'luxon';
 import { useCallback, useEffect, useMemo, useState } from 'preact/hooks';
-import { Button, Checkbox, Divider, Modal } from 'react-daisyui';
+import { Button, Checkbox, Modal } from 'react-daisyui';
 import { Controller, useForm } from 'react-hook-form';
 import { toast } from 'react-toastify';
 import { Asset } from 'stellar-sdk';
-import { useGlobalState } from '../../GlobalStateProvider';
-import { useNodeInfoState } from '../../NodeInfoProvider';
-import LabelledInputField from '../../components/LabelledInputField';
-import { CopyableAddress, PublicKey } from '../../components/PublicKey';
-import { AssetSelector, VaultSelector } from '../../components/Selector';
-import TransferCountdown from '../../components/TransferCountdown';
-import OpenWallet from '../../components/Wallet';
-import { calculateDeadline, convertCurrencyToStellarAsset, deriveShortenedRequestId } from '../../helpers/spacewalk';
-import { convertRawHexKeyToPublicKey, isCompatibleStellarAmount, stringifyStellarAsset } from '../../helpers/stellar';
-import { getErrors, getEventBySectionAndMethod } from '../../helpers/substrate';
-import { useFeePallet } from '../../hooks/spacewalk/fee';
-import { RichIssueRequest, useIssuePallet } from '../../hooks/spacewalk/issue';
-import { useSecurityPallet } from '../../hooks/spacewalk/security';
-import { ExtendedRegistryVault, useVaultRegistryPallet } from '../../hooks/spacewalk/vaultRegistry';
-import { decimalToStellarNative, nativeStellarToDecimal, nativeToDecimal } from '../../shared/parseNumbers';
+import { useGlobalState } from '../../../GlobalStateProvider';
+import { useNodeInfoState } from '../../../NodeInfoProvider';
+import LabelledInputField from '../../../components/LabelledInputField';
+import { CopyableAddress, PublicKey } from '../../../components/PublicKey';
+import { AssetSelector, VaultSelector } from '../../../components/Selector';
+import OpenWallet from '../../../components/Wallet';
+import { convertCurrencyToStellarAsset } from '../../../helpers/spacewalk';
+import {
+  StellarPublicKeyPattern,
+  convertRawHexKeyToPublicKey,
+  isCompatibleStellarAmount,
+  isPublicKey,
+  stringifyStellarAsset,
+} from '../../../helpers/stellar';
+import { getErrors, getEventBySectionAndMethod } from '../../../helpers/substrate';
+import { useFeePallet } from '../../../hooks/spacewalk/fee';
+import { RichRedeemRequest, useRedeemPallet } from '../../../hooks/spacewalk/redeem';
+import { ExtendedRegistryVault, useVaultRegistryPallet } from '../../../hooks/spacewalk/vaultRegistry';
+import { decimalToStellarNative, nativeStellarToDecimal, nativeToDecimal } from '../../../shared/parseNumbers';
 
 interface FeeBoxProps {
   bridgedAsset?: Asset;
@@ -31,16 +32,15 @@ interface FeeBoxProps {
   amountNative: Big;
   extrinsic?: SubmittableExtrinsic;
   network: string;
-  wrappedCurrencySuffix?: string;
   nativeCurrency: string;
+  wrappedCurrencySuffix?: string;
 }
 
 function FeeBox(props: FeeBoxProps): JSX.Element {
-  const { bridgedAsset, extrinsic, network, wrappedCurrencySuffix, nativeCurrency } = props;
-
+  const { bridgedAsset, extrinsic, nativeCurrency, wrappedCurrencySuffix, network } = props;
   const amount = props.amountNative;
 
-  const wrappedCurrencyName = bridgedAsset ? bridgedAsset.getCode() + (wrappedCurrencySuffix || '') : '';
+  const wrappedCurrencyName = bridgedAsset ? (wrappedCurrencySuffix || '') + bridgedAsset.getCode() : '';
 
   const { getFees, getTransactionFee } = useFeePallet();
   const fees = getFees();
@@ -58,11 +58,7 @@ function FeeBox(props: FeeBoxProps): JSX.Element {
   }, [extrinsic, getTransactionFee, setTransactionFee]);
 
   const bridgeFee = useMemo(() => {
-    return nativeStellarToDecimal(amount.mul(fees.issueFee));
-  }, [amount, fees]);
-
-  const griefingCollateral = useMemo(() => {
-    return nativeStellarToDecimal(amount.mul(fees.issueGriefingCollateral));
+    return nativeStellarToDecimal(amount.mul(fees.redeemFee));
   }, [amount, fees]);
 
   const totalAmount = useMemo(() => {
@@ -74,7 +70,7 @@ function FeeBox(props: FeeBoxProps): JSX.Element {
   }, [amount, bridgeFee]);
 
   return (
-    <div className="shadow bg-base-100 rounded-lg p-4 my-4 flex flex-col text-sm xs:text-base">
+    <div className="shadow bg-base-100 rounded-lg p-4 my-4 flex flex-col">
       <div className="flex justify-between">
         <span>To {network}</span>
         <span>
@@ -83,19 +79,13 @@ function FeeBox(props: FeeBoxProps): JSX.Element {
       </div>
       <div className="flex justify-between mt-2">
         <span>Bridge Fee</span>
-        <span className="text-right">
+        <span>
           {bridgeFee.toString()} {bridgedAsset?.getCode()}
         </span>
       </div>
       <div className="flex justify-between mt-2">
-        <span>Security Deposit</span>
-        <span className="text-right">
-          {griefingCollateral.toString()} {nativeCurrency}
-        </span>
-      </div>
-      <div className="flex justify-between mt-2">
         <span>Transaction Fee</span>
-        <span className="text-right">
+        <span>
           {transactionFee.toFixed(12)} {nativeCurrency}
         </span>
       </div>
@@ -104,164 +94,112 @@ function FeeBox(props: FeeBoxProps): JSX.Element {
 }
 
 interface ConfirmationDialogProps {
-  issueRequest: RichIssueRequest | undefined;
+  redeemRequest: RichRedeemRequest | undefined;
   onClose: () => void;
   visible: boolean;
 }
 
 function ConfirmationDialog(props: ConfirmationDialogProps): JSX.Element {
-  const { issueRequest, visible, onClose } = props;
+  const { redeemRequest, visible, onClose } = props;
 
-  const { subscribeActiveBlockNumber } = useSecurityPallet();
-  const [activeBlockNumber, setActiveBlockNumber] = useState<number>(0);
-  const [_, setRemainingDurationString] = useState<string>('');
-
-  const totalAmount = useMemo(
-    () =>
-      issueRequest
-        ? nativeStellarToDecimal(issueRequest.request.amount.add(issueRequest.request.fee).toString()).toString()
-        : '',
-    [issueRequest],
-  );
-
-  const asset = useMemo(() => {
-    const currency = issueRequest?.request.asset;
-    return currency && convertCurrencyToStellarAsset(currency);
-  }, [issueRequest?.request.asset]);
+  const totalAmount = redeemRequest ? nativeStellarToDecimal(redeemRequest.request.amount.toString()).toString() : '';
+  const currency = redeemRequest?.request.asset;
+  const asset = currency && convertCurrencyToStellarAsset(currency);
 
   const destination = useMemo(() => {
-    const rawDestinationAddress = issueRequest?.request.stellarAddress;
+    const rawDestinationAddress = redeemRequest?.request.stellarAddress;
     return rawDestinationAddress ? convertRawHexKeyToPublicKey(rawDestinationAddress.toHex()).publicKey() : '';
-  }, [issueRequest?.request.stellarAddress]);
-
-  const expectedStellarMemo = useMemo(() => {
-    if (!issueRequest) {
-      return '';
-    }
-    // For issue requests we use a shorter identifier for the memo
-    return deriveShortenedRequestId(issueRequest.id);
-  }, [issueRequest]);
-
-  useEffect(() => {
-    let unsub: VoidFn = () => undefined;
-    subscribeActiveBlockNumber((blockNumber) => {
-      setActiveBlockNumber(blockNumber);
-    }).then((u) => (unsub = u));
-
-    return unsub;
-  }, [subscribeActiveBlockNumber]);
-
-  const deadline = useMemo(() => {
-    const openTime = issueRequest?.request.opentime.toNumber() || 0;
-    const period = issueRequest?.request.period.toNumber() || 0;
-    const end = calculateDeadline(activeBlockNumber, openTime, period, 12);
-
-    return end;
-  }, [activeBlockNumber, issueRequest]);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      const newDeadlineString = deadline.diff(DateTime.now()).toFormat('hh:mm:ss');
-      setRemainingDurationString(newDeadlineString);
-    });
-
-    return () => clearInterval(interval);
-  }, [deadline]);
+  }, [redeemRequest?.request.stellarAddress]);
 
   return (
     <Modal open={visible}>
-      <Modal.Header className="font-bold">Deposit</Modal.Header>
+      <Modal.Header className="font-bold">Back to Stellar</Modal.Header>
       <Button color="ghost" size="md" shape="circle" className="absolute right-4 top-4" onClick={onClose}>
         ✕
       </Button>
       <Modal.Body>
         <div className="text-center">
           <div className="text-xl">
-            Send {totalAmount} {asset?.getCode()}
+            You will receive {totalAmount} {asset?.getCode()}
           </div>
-          <div className="text-sm">
-            {asset && asset.getIssuer() && (
-              <>
-                issued by <PublicKey variant="short" publicKey={asset?.getIssuer()} />
-              </>
-            )}
-          </div>
-          <div className="text mt-4">With the text memo</div>
-          {issueRequest && <CopyableAddress variant="short" publicKey={expectedStellarMemo} />}
-          <div className="text mt-4">In a single transaction to</div>
-          <CopyableAddress variant="short" publicKey={destination} />
-          <div className="mt-4">Within {issueRequest && <TransferCountdown request={issueRequest?.request} />}</div>
-        </div>
-        <Divider />
-        <div>
-          {asset?.getAssetType() !== 'native' && (
-            <div className="text-sm">
-              Warning: Make sure that the {asset?.code} you are sending are issued by the correct issuer.
-            </div>
+          {asset && asset.getIssuer() && (
+            <>
+              issued by <PublicKey variant="short" publicKey={asset?.getIssuer()} />
+            </>
           )}
+          <div className="text-sm text-secondary mt-4">Your request is being processed</div>
         </div>
-        <div className="text-sm mt-4">
-          Note: If you have already made the payment, please wait for a few minutes for it to be confirmed.
+        <div className="mt-6 text-secondary">
+          <div className="flex items-center justify-between">
+            <span>Stellar destination address</span>
+            <CopyableAddress variant="short" publicKey={destination} />
+          </div>
+          <div className="text-sm mt-2">
+            We will inform you when the PEN payment is executed. This typically takes only a few minutes but may
+            sometimes take up to 6 hours.
+          </div>
         </div>
       </Modal.Body>
 
       <Modal.Actions className="justify-center">
         <Button color="primary" onClick={onClose}>
-          I have made the payment
+          View Progress
         </Button>
       </Modal.Actions>
     </Modal>
   );
 }
 
-interface IssueFormInputs {
+interface RedeemFormInputs {
   amount: string;
+  stellarAddress: string;
 }
 
-interface IssueProps {
+interface RedeemProps {
   network: string;
   wrappedCurrencySuffix: string;
   nativeCurrency: string;
 }
 
-function Issue(props: IssueProps): JSX.Element {
-  const { network, wrappedCurrencySuffix, nativeCurrency } = props;
-
+function Redeem(props: RedeemProps): JSX.Element {
   const [selectedVault, setSelectedVault] = useState<ExtendedRegistryVault>();
   const [selectedAsset, setSelectedAsset] = useState<Asset>();
-  const [manualVaultSelection, setManualVaultSelection] = useState(false);
   const [confirmationDialogVisible, setConfirmationDialogVisible] = useState(false);
   const [submissionPending, setSubmissionPending] = useState(false);
-  const [submittedIssueRequest, setSubmittedIssueRequest] = useState<RichIssueRequest | undefined>(undefined);
+  const [submittedRedeemRequest, setSubmittedRedeemRequest] = useState<RichRedeemRequest | undefined>(undefined);
+  const [manualVaultSelection, setManualVaultSelection] = useState(false);
   const [vaults, setExtendedVaults] = useState<ExtendedRegistryVault[]>();
 
-  const { createIssueRequestExtrinsic, getIssueRequest } = useIssuePallet();
-  const { getVaults, getVaultsWithIssuableTokens } = useVaultRegistryPallet();
+  const { createRedeemRequestExtrinsic, getRedeemRequest } = useRedeemPallet();
+  const { getVaults, getVaultsWithRedeemableTokens } = useVaultRegistryPallet();
   const { walletAccount, dAppName } = useGlobalState();
-
   const { api } = useNodeInfoState().state;
 
-  const { control, handleSubmit, watch } = useForm<IssueFormInputs>({
+  const { control, handleSubmit, watch } = useForm<RedeemFormInputs>({
     defaultValues: {
       amount: '0',
+      stellarAddress: '',
     },
   });
 
+  const { wrappedCurrencySuffix, nativeCurrency } = props;
+
   // We watch the amount because we need to re-render the FeeBox constantly
   const amount = watch('amount');
+  const stellarAddress = watch('stellarAddress');
 
   useEffect(() => {
     const combinedVaults: ExtendedRegistryVault[] = [];
-    getVaultsWithIssuableTokens().then((vaultsWithIssuableTokens) => {
+    getVaultsWithRedeemableTokens().then((vaultsWithRedeemableTokens) => {
       getVaults().forEach((vaultFromRegistry) => {
-        const found = vaultsWithIssuableTokens?.find(([id, _]) => id.eq(vaultFromRegistry.id));
+        const found = vaultsWithRedeemableTokens?.find(([id, _]) => id.eq(vaultFromRegistry.id));
         const extended: ExtendedRegistryVault = vaultFromRegistry;
-        extended.issuableTokens = found ? found[1] : undefined;
+        extended.redeemableTokens = found ? found[1] : undefined;
         combinedVaults.push(extended);
       });
       setExtendedVaults(combinedVaults);
     });
-  }, [getVaults, setExtendedVaults, getVaultsWithIssuableTokens]);
+  }, [getVaults, setExtendedVaults, getVaultsWithRedeemableTokens]);
 
   // The amount represented in the units of the native currency (as integer)
   const amountNative = useMemo(() => {
@@ -270,6 +208,7 @@ function Issue(props: IssueProps): JSX.Element {
 
   const wrappedAssets = useMemo(() => {
     if (!vaults) return;
+
     const assets = vaults
       .map((vault) => {
         const currency = vault.id.currencies.wrapped;
@@ -307,34 +246,34 @@ function Issue(props: IssueProps): JSX.Element {
         }
       } else {
         // If the user manually selected a vault, but it's not available anymore, we reset the selection
-        if (selectedVault && !vaultsForCurrency.includes(selectedVault) && vaultsForCurrency.length > 0) {
+        if (selectedVault && vaultsForCurrency.length > 0 && !vaultsForCurrency.includes(selectedVault)) {
           setSelectedVault(vaultsForCurrency[0]);
         }
       }
     }
   }, [manualVaultSelection, selectedAsset, selectedVault, vaultsForCurrency, wrappedAssets]);
 
-  const requestIssueExtrinsic = useMemo(() => {
-    if (!selectedVault || !api) {
+  const requestRedeemExtrinsic = useMemo(() => {
+    if (!selectedVault || !api || !stellarAddress || !isPublicKey(stellarAddress)) {
       return undefined;
     }
 
-    return createIssueRequestExtrinsic(amountNative.toString(), selectedVault.id);
-  }, [amountNative, api, createIssueRequestExtrinsic, selectedVault]);
+    return createRedeemRequestExtrinsic(amountNative.toString(), stellarAddress, selectedVault.id);
+  }, [amountNative, api, createRedeemRequestExtrinsic, selectedVault, stellarAddress]);
 
-  const submitRequestIssueExtrinsic = useCallback(() => {
-    if (!requestIssueExtrinsic || !api || !selectedVault) {
+  const submitRequestRedeemExtrinsic = useCallback(() => {
+    if (!requestRedeemExtrinsic || !api || !selectedVault) {
       return;
     }
 
     if (!walletAccount) {
-      toast('No wallet account selected', { type: 'error' });
       return;
     }
 
     setSubmissionPending(true);
 
-    requestIssueExtrinsic
+    requestRedeemExtrinsic
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
       .signAndSend(walletAccount.address, { signer: walletAccount.signer as any }, (result) => {
         const { status, events } = result;
 
@@ -346,15 +285,16 @@ function Issue(props: IssueProps): JSX.Element {
             toast(errorMessage, { type: 'error' });
           }
         } else if (status.isFinalized) {
-          const requestIssueEvents = getEventBySectionAndMethod(events, 'issue', 'RequestIssue');
+          const requestRedeemEvents = getEventBySectionAndMethod(events, 'redeem', 'RequestRedeem');
 
           // We only expect one event but loop over all of them just in case
-          for (const requestIssueEvent of requestIssueEvents) {
+          for (const requestRedeemEvent of requestRedeemEvents) {
             // We do not have a proper type for this event, so we have to cast it to any
-            const issueId = (requestIssueEvent.data as any).issueId;
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            const redeemId = (requestRedeemEvent.data as any).redeemId;
 
-            getIssueRequest(issueId).then((issueRequest) => {
-              setSubmittedIssueRequest(issueRequest);
+            getRedeemRequest(redeemId).then((redeemRequest) => {
+              setSubmittedRedeemRequest(redeemRequest);
             });
           }
 
@@ -367,23 +307,23 @@ function Issue(props: IssueProps): JSX.Element {
       })
       .catch((error) => {
         console.error('Transaction submission failed', error);
-        toast('Transaction submission failed: ' + error.toString(), {
+        toast('Transaction submission failed:' + error.toString(), {
           type: 'error',
         });
         setSubmissionPending(false);
       });
-  }, [api, getIssueRequest, requestIssueExtrinsic, selectedVault, walletAccount]);
+  }, [api, getRedeemRequest, requestRedeemExtrinsic, selectedVault, walletAccount]);
 
   return (
-    <div className="flex items-center justify-center h-full space-walk py-4">
+    <div className="flex items-center justify-center h-full space-walk py-4 w-full">
       <ConfirmationDialog
-        issueRequest={submittedIssueRequest}
+        redeemRequest={submittedRedeemRequest}
         visible={confirmationDialogVisible}
         onClose={() => setConfirmationDialogVisible(false)}
       />
       <div className="w-full">
         <form className="px-5 flex flex-col">
-          <div className="flex flex-col xs:flex-row xs:items-center gap-1">
+          <div className="flex items-center">
             <Controller
               control={control}
               rules={{
@@ -409,7 +349,7 @@ function Issue(props: IssueProps): JSX.Element {
                 <LabelledInputField
                   autoSelect
                   error={error?.message}
-                  label="From Stellar"
+                  label="From Amplitude"
                   type="number"
                   min="0"
                   step="any"
@@ -423,8 +363,10 @@ function Issue(props: IssueProps): JSX.Element {
                 />
               )}
             />
+            <div className="px-1" />
             {wrappedAssets && (
               <AssetSelector
+                assetSuffix={wrappedCurrencySuffix}
                 selectedAsset={selectedAsset}
                 assets={wrappedAssets}
                 onChange={setSelectedAsset}
@@ -433,11 +375,12 @@ function Issue(props: IssueProps): JSX.Element {
             )}
           </div>
           <div className="flex align-center mt-4">
-            <span className="text-sm">{`Max issuable: ${nativeToDecimal(
-              selectedVault?.issuableTokens?.toString() || 0,
+            <span className="text-sm">{`Max redeemable: ${nativeToDecimal(
+              selectedVault?.redeemableTokens?.toString() || 0,
             ).toFixed(2)} 
               ${selectedAsset?.code}`}</span>
           </div>
+
           <div className="flex align-center mt-4">
             <Checkbox
               size="sm"
@@ -455,15 +398,35 @@ function Issue(props: IssueProps): JSX.Element {
               vaults={vaultsForCurrency}
               onChange={setSelectedVault}
               selectedVault={selectedVault}
-              showMaxTokensFor="issuableTokens"
+              showMaxTokensFor="redeemableTokens"
             />
           )}
+          <Controller
+            control={control}
+            rules={{
+              required: 'Stellar address is required',
+              pattern: {
+                value: StellarPublicKeyPattern,
+                message: 'Stellar address is invalid',
+              },
+            }}
+            render={({ field, fieldState: { error } }) => (
+              <LabelledInputField
+                label="Stellar Address"
+                error={error?.message}
+                placeholder="Enter target Stellar address"
+                type="text"
+                {...field}
+                style={{ marginTop: 8 }}
+              />
+            )}
+            name="stellarAddress"
+          />
           <FeeBox
             amountNative={amountNative}
             bridgedAsset={selectedAsset}
-            extrinsic={requestIssueExtrinsic}
-            network={network}
-            wrappedCurrencySuffix={wrappedCurrencySuffix}
+            extrinsic={requestRedeemExtrinsic}
+            network="Stellar"
             nativeCurrency={nativeCurrency}
           />
           {walletAccount ? (
@@ -471,7 +434,7 @@ function Issue(props: IssueProps): JSX.Element {
               className="w-full"
               color="primary"
               loading={submissionPending}
-              onClick={handleSubmit(submitRequestIssueExtrinsic)}
+              onSubmit={handleSubmit(submitRequestRedeemExtrinsic)}
               type="button"
             >
               Bridge
@@ -485,4 +448,4 @@ function Issue(props: IssueProps): JSX.Element {
   );
 }
 
-export default Issue;
+export default Redeem;
